refactor: Theme provider
